% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\docType{data}
\name{anndata_struct_names}
\alias{anndata_struct_names}
\title{Read anndata slots}
\format{
An object of class \code{character} of length 8.
}
\usage{
anndata_struct_names
}
\arguments{
\item{spec}{file spec}

\item{path}{Path to yaml file}
}
\value{
tibble with file slots
}
\description{
Read anndata slots
}
\examples{
\dontrun{
read_anndata_slots(spec, 'path/to/yaml')
}
}
\keyword{datasets}
