% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_project_root.R
\name{find_project_root}
\alias{find_project_root}
\title{Find the root of a Viash project}
\usage{
find_project_root(path = ".")
}
\arguments{
\item{path}{Path to a file or directory}
}
\value{
The path to the root of the Viash project, or NULL if not found
}
\description{
This function will recursively search for a \verb{_viash.yaml} file
in the parent directories of the given path.
}
\examples{
\dontrun{
find_project_root("/path/to/project/subdir")
}
}
