% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openproblems-package.R
\docType{package}
\name{openproblems-package}
\alias{openproblems}
\alias{openproblems-package}
\title{openproblems: OpenProblems Core}
\description{
OpenProblems core R helper functions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://openproblems.bio/documentation}
  \item \url{https://github.com/openproblems-bio/core}
  \item Report bugs at \url{https://github.com/openproblems-bio/core/issues}
}

}
\author{
\strong{Maintainer}: Robrecht Cannoodt \email{rcannood@gmail.com} (\href{https://orcid.org/0000-0003-3641-729X}{ORCID}) (rcannood)

}
\keyword{internal}
