% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_api_files.R
\name{read_api_files}
\alias{read_api_files}
\title{Read the api files in a task}
\usage{
read_api_files(path)
}
\arguments{
\item{path}{Path to a src directory}
}
\value{
A list with the api info
}
\description{
Read the api files in a task
}
\examples{
\dontrun{
read_api_files("src")
}
}
