% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_api.R
\name{read_comp_args}
\alias{read_comp_args}
\title{read component arguments
#'}
\usage{
read_comp_args(spec_yaml, path)
}
\arguments{
\item{spec_yaml}{comp yaml}

\item{path}{path to comp yaml}
}
\value{
a df with component arguments
}
\description{
read component arguments
#'
}
\examples{
\dontrun{
read_comp_args(spec_yaml, "src/path/to/comp.yaml")
}
}
