% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_api.R
\name{read_comp_info}
\alias{read_comp_info}
\title{read component spec
#'}
\usage{
read_comp_info(spec_yaml, path)
}
\arguments{
\item{spec_yaml}{comp yaml}

\item{path}{path to comp yaml}
}
\value{
A tibble with component info
}
\description{
read component spec
#'
}
\examples{
\dontrun{
read_comp_info(spec_yaml, "src/path/to/comp.yaml")
}
}
