% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_file_spec}
\alias{read_file_spec}
\title{Read file specification}
\usage{
read_file_spec(path)
}
\arguments{
\item{path}{Path to yaml file}
}
\value{
list with file info and slots
}
\description{
Read file specification
}
\examples{
\dontrun{
read_file_spec('path/to/yaml')
}
}
