% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_nested_yaml.R
\name{read_nested_yaml}
\alias{read_nested_yaml}
\title{Read a nested YAML}
\usage{
read_nested_yaml(path, project_path = find_project_root(path))
}
\arguments{
\item{path}{Path to a YAML}

\item{project_path}{Path to the Viash project root}
}
\value{
A list with the merged YAML
}
\description{
If the YAML contains a "\strong{merge}" key anywhere in the yaml,
the path specified in that YAML will be read and the two
lists will be merged. This is a recursive procedure.
}
\examples{
\dontrun{
read_nested_yaml("src/path/to/config.vsh.yaml")
}
}
