% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_file.R
\name{read_tabular_columns}
\alias{read_tabular_columns}
\title{Read columns from tabular file}
\usage{
read_tabular_columns(spec, path)
}
\arguments{
\item{spec}{file spec}

\item{path}{Path to yaml file}
}
\value{
tibble with columns
}
\description{
Read columns from tabular file
}
\examples{
\dontrun{
read_tabular_columns(spec, 'path/to/yaml')
}
}
