% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_viash_config.R
\name{read_viash_config}
\alias{read_viash_config}
\title{Process a viash config file}
\usage{
read_viash_config(
  target_config_path,
  project_root_dir = find_project_root(target_config_path)
)
}
\arguments{
\item{target_config_path}{Path to a Viash config file in the target directory.
Example: \code{target/executable/path/to/.config.vsh.yaml}.}

\item{project_root_dir}{Path to the root of the Viash project. If not provided,
it will be determined automatically based on the target config path.}
}
\value{
The processed config as a named list.
}
\description{
This function will process a viash config file in the following ways:
}
\details{
\itemize{
\item Upgrade the config from viash < 0.9 to viash >= 0.9 format.
\item Add 'full_name' field to the config, which is the namespace/name of the config.
\item Add 'all_arguments' field to the config, which is a flattened list of all arguments
in the config, with an additional 'clean_name' field that is the argument name without
any leading dashes.
}
}
