% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_components.R
\name{render_component}
\alias{render_component}
\title{Render component section}
\usage{
render_component(spec)
}
\arguments{
\item{spec}{file spec}
}
\value{
string
}
\description{
Render component section
}
\examples{
\dontrun{
render_component(spec)
}
}
