% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{render_example}
\alias{render_example}
\title{Render an example yaml based on a JSON schema}
\usage{
render_example(json_schema)
}
\arguments{
\item{json_schema}{json schema}
}
\value{
string
}
\description{
Render an example yaml based on a JSON schema
}
\examples{
\dontrun{
render_example('path/to/json')
}
}
