% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/task_graph.R
\name{render_task_graph}
\alias{render_task_graph}
\title{Render the task graph
#'}
\usage{
render_task_graph(task_api, root = .task_graph_get_root(task_api))
}
\arguments{
\item{task_api}{the task api filepath}

\item{root}{root of the task graph}
}
\value{
igraph
}
\description{
Render the task graph
#'
}
\examples{
\dontrun{
render_task_graph(task_api)
}
}
